Refactor: Organize Rust code into multiple files
I've divided your codebase in `src/main.rs` into logical modules:
- `src/structs.rs`: For data structure definitions (ButtonInfo, OutgoingKafkaMessage, IncomingCallbackMessage).
- `src/telegram_handlers.rs`: For Telegram message and callback query handlers (message_handler, callback_query_handler).
- `src/utils.rs`: For utility functions (create_markup, prepare_incoming_callback_message).
- `src/kafka_processing.rs`: For the Kafka consumer loop that sends messages to Telegram.

I've updated `main.rs` to declare these modules and use their public functions/structs.

NOTE: Compilation failed after these changes.
Your build environment uses rustc 1.75.0, which is incompatible with
several direct and transitive dependencies of the project (even when
attempting to use older versions of 'teloxide' and Rust edition "2021").
Your project will require a newer Rust compiler (likely 1.82+ or newer)
to build successfully. The refactoring itself is structurally sound.
